+++ src/lib.rs
@@ -140,6 +140,7 @@ pub mod options;
 pub mod profile;
 pub mod screen;
 pub mod scroll;
+pub mod tablist;
 pub mod tabs;
 pub mod text;
 pub mod upload;